Updated gitignore and setup

+++ e-cart/setup.sh
@@ -46,12 +46,61 @@ for file in $(ls models); do
         sed -i s/'});'/'\ \ },{timestamps: false, freezeTableName: true,});'/g models/${file}
 done
 
-# 4. Install express
+# 4.1 Save existing files
+mkdir tmp
+if [ -f app.js ]; then
+        cp app.js tmp/app.js
+fi
+
+if [ -f routes/index.js ]; then
+        cp routes/index.js tmp/index.js
+fi
+
+if [ -f views/index.jade ]; then
+        cp views/index.jade tmp/index.jade
+fi
+
+if [ -f views/layout.jade ]; then
+        cp views/layout.jade tmp/layout.jade
+fi
+
+if [ -f public/stylesheets/containers.css ]; then
+	cp public/stylesheets/containers.css tmp/containers.css
+fi
+# 4.2 Deploy express
 yes | express --sessions --css stylus
+# 4.3 Install deps
 npm install
-# 4.1 Deploy Files
-cp install/index.js routes/index.js
-cp install/index.jade views/index.jade
-cp install/layout.jade views/layout.jade
-cp install/containers.css public/stylesheets/containers.css
-cp install/app.js ./app.js
+# 4.4 Deploy/Restore Files
+if [ -f tmp/app.js ]; then
+        mv tmp/app.js ./app.js
+else
+        cp install/app.js ./app.js
+fi
+
+if [ -f tmp/index.js ]; then
+        mv tmp/index.js routes/index.js
+else
+        cp install/index.js routes/index.js
+fi
+
+if [ -f tmp/index.jade ]; then
+        mv tmp/index.jade views/index.jade
+else
+        cp install/index.jade views/index.jade
+fi
+
+if [ -f tmp/layout.jade ]; then
+        mv tmp/layout.jade views/layout.jade
+else
+        cp install/layout.jade views/layout.jade
+fi
+
+if [ -f tmp/containers.css ]; then
+	mv tmp/containers.css public/stylesheets/containers.css
+else
+	cp install/containers.css public/stylesheets/containers.css
+fi
+
+rmdir tmp
+
